@@ -1,45 +1,39 @@
 {
   "name": "electron-react-webpack-setup",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "electron react webpack setup",
   "main": "index.js",
   "scripts": {
-    "electron-rebuild": "./node_modules/.bin/electron-rebuild",
     "start": "npm run build; node ./node_modules/.bin/electron .",
-    "build": "node ./node_modules/.bin/webpack --config webpack.config.js",
-    "watch": "node ./node_modules/.bin/webpack --config webpack.config.js --progress --colors --watch"
+    "build": "node ./node_modules/.bin/webpack --config webpack.dev.js",
+    "watch": "node ./node_modules/.bin/webpack --config webpack.dev.js --progress --colors --watch"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/demeter-macik/electron-react-webpack-setup.git"
   },
-  "dependencies": {
-    "config": "^1.26.2",
-    "debug": "^3.0.0",
-    "knex": "^0.13.0",
-    "material-ui": "^0.19.0",
-    "sqlite3": "^3.1.12"
-  },
   "devDependencies": {
-    "eslint-plugin-react": "^7.2.1",
-    "babel-cli": "^6.24.1",
-    "babel-loader": "^7.1.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-3": "^6.24.1",
-    "electron": "^1.7.5",
-    "electron-builder": "^19.31.1",
-    "electron-rebuild": "^1.6.0",
-    "electron-reload": "^1.2.2",
-    "eslint": "^4.4.1",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "webpack": "^3.5.4"
+    "@babel/core": "^7.1.2",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/preset-react": "^7.0.0",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.4",
+    "electron": "^3.0.6",
+    "electron-reload": "^1.2.5",
+    "eslint": "^5.8.0",
+    "eslint-plugin-react": "^7.11.1",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0",
+    "webpack": "^4.23.1",
+    "webpack-cli": "^3.1.2"
   },
-  "author": "",
-  "license": "ISC",
+  "author": "Demeter Macik",
+  "license": "GPL-3.0",
   "bugs": {
     "url": "https://github.com/demeter-macik/electron-react-webpack-setup/issues"
   },
-  "homepage": "https://github.com/demeter-macik/electron-react-webpack-setup#readme"
+  "homepage": "https://github.com/demeter-macik/electron-react-webpack-setup#readme",
+  "dependencies": {
+    "material-ui": "^0.20.2"
+  }
 }
