@@ -4,17 +4,21 @@ const SRC_DIR = path.resolve(__dirname, 'src');
 const DIST_DIR = path.resolve(__dirname, 'dist');
 
 module.exports = {
+    mode: 'development',
     entry: './src/main.js',
     output: {
         path: DIST_DIR,
         filename: 'bundle.js',
     },
     module: {
-        loaders: [
+        rules: [
             {
                 test: /\.js$/,
                 include: SRC_DIR,
-                loader: 'babel-loader',
+                exclude: /node_modules/,
+                use: {
+                    loader: "babel-loader"
+                }
             }
         ]
     }
